+++ sample_codes/zk_example_individual_steps.py
@@ -32,5 +32,7 @@ if zk.project_menu.exists():
     print(zk.project_menu.read_text(encoding="utf-8"))
     print(f'Total {menu.sum_of_items()} files generated.')
 
-print(f'Elapsed time for each step (sec) = {zk.takt_time}')
-print(f'Total elapsed time (sec) = {sum(zk.takt_time)}')
+print(f'Elapsed time for each step (sec):')
+for step, elapsed_time in zk.takt_time.items():
+    print(f"    {step}: {elapsed_time}")
+print(f'Total elapsed time (sec) = {sum(zk.takt_time.values())}')