+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Zoltraak Klein Changelog
 
+## [1.0.1] - 2024-08-21
+### Changed
+- For architect "epub" and "epub_picture", corrected bug of recording file name to the menu.
+- `takt_time` changed from list to dictionary.
+
 ## [1.0.0] - 2024-08-19
 ### Changed
 - Formal release of Zoltraak Klein